Moved database and data file serializers into their own files.
